Add an import option to force TextureArrays as normal maps for RGTC compression
this is a redo of the pull request : https://github.com/godotengine/godot/pull/52058

i'm a noob on github and i deleted it. 

i have added the option to chose normal map for channel pack. this will force the to use red and green channels for the texture only. the option is added meaning it will not conflict with the existing settings.

i have made this change as a response to my own proposal